refactored some functions and constants from main_manual.py to connector/blob.py and the configuration file bug fix: main_manual.py whitespace stripping from links Renamed log variables from manual run to Azure Function run Uploaded CSVs append to monthly CSVs (like 2024-10.csv)
